bot_framework/requirements.txt
APScheduler==3.6.3
arrow==1.2.2
asgiref==3.4.1
aspose-cells==21.12.1
async-generator==1.10
attrs==21.2.0
azure-core==1.22.1
azure-storage-blob==12.11.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
covid==2.0.7
cryptography==3.4.8
cycler==0.10.0
DateTime==4.4
defusedxml==0.7.1
Django==3.0.8
django-allauth==0.45.0
django-cors-headers==3.7.0
django-filter==21.1
django-formset-js-improved==0.5.0.2
django-jquery-js==3.1.1
django-jsonforms==1.1.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
glob2==0.7
google-api-core==1.31.3
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.53.0
gunicorn==20.0.4
h11==0.13.0
httplib2==0.20.1
idna==2.8
importlib-metadata==4.8.1
isodate==0.6.1
JPype1==1.3.0
jsonschema==4.0.1
linecache2==1.0.0
lxml==4.8.0
Markdown==3.3.4
marshmallow==3.14.0
marshmallow-jsonschema==0.12.0
mongoengine==0.22.1
mongomock==3.21.0
msrest==0.6.21
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.0
pandas==1.3.5
pathlib==1.0.1
Pillow==9.0.1
pip==21.0.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.12.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.7.5
python-docx==0.8.1
python-gitlab==2.5.0
python-pptx==0.6.21
python-telegram-bot==13.7
python-time==0.3.0
python3-openid==3.2.0
pytz==2019.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
rstr==2.2.6
selenium==4.1.5
sentinels==1.0.0
setuptools==53.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
telegram==0.0.1
tornado==6.1
tqdm==4.63.1
traceback2==1.4.0
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
tzdata==2021.2.post0
tzlocal==3.0
uritemplate==3.0.1
urllib3==1.26.9
vobject==0.9.6.1
webdriver-manager==3.5.4
wheel==0.36.2
wsproto==1.1.0
XlsxWriter==3.0.3
xmltodict==0.12.0
zipp==3.6.0
zope.interface==5.4.0
